Added a countdown to the tapped code, also added level mechanics, so far it goes all the way to level 24, also known as 2^24 codes to tap

+++ app/src/main/java/com/example/ergo/incremental/core_mechanics/Game.java
@@ -13,13 +13,13 @@ public class Game {
     public static int currentLevel = GameValues.STARTING_LEVEL;
 
     // FIXME: Duplicate code execution, is there a cleaner way to do this?
-    public static double codeToMake = Math.pow(GameValues.CODE_A_CREER_DE_BASE, currentLevel);
+    public static int codeToMake = (int)Math.pow(GameValues.CODE_A_CREER_DE_BASE, currentLevel);
 
     public static void calculateCodeToMake(){
-        codeToMake = Math.pow(GameValues.CODE_A_CREER_DE_BASE, currentLevel);
+        codeToMake = (int)Math.pow(GameValues.CODE_A_CREER_DE_BASE, currentLevel);
     }
 
-    public static  void levelUp() {
+    public static void levelUp() {
         if(currentLevel < GameValues.MAX_NIVEAU){
             currentLevel++;
             calculateCodeToMake();
